Tweak some models in the datasets app
This adds some (optional) identifiers (for use when importing datasets
with a script for example) and increases the precision of the
`DecimalField` in `DataPoint`.

Moreover, it accounts for the fact that a metric can belong to more than
one `DatasetSchema` be removing the FK to that model and replacing it
with a many-to-many relationship.

This makes some changes to the REST API necessary. For now I just
removed everything from there that has anything to do with the old
`schema` field of `DatasetMetric`. I'm not sure if this breaks anything.