Implement bulk editing for data points
This includes a fix for inconsistent serialization of UUID fields.
Depending on whether we refer to an UUID via a `UUIDField` or a
`SlugRelatedField`, the value would be either a string or a `UUID`
object. This causes us to fail finding objects by UUID.

To fix this, this commit introduces a subclass of
`SlugRelatedField` called `UuidSlugRelatedField`, which should be used
whenever we refer to a related object by UUID. The new field type takes
care of serializing UUIDs to strings.